UTL-387 Upgrade.
The main issue was the import of the Node 'url' module. Webpack 5 stops importing polyfills for these things and they don't exist in the browser.

We only used it for IE11 support, so I've removed all that and the import, as well as some minor tidying up.

Added the 'files' property to package.json, so that we don't include extra junk in the published package.